Move methods providers to class providers.
