+++ tests/Provider/QueryProvider.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Db\Tests\Provider;
+
+final class QueryProvider extends AbstractQueryProvider
+{
+}
